requirements.txt
django==3.1.6
djangorestframework
django-rest-knox
django-cors-headers